+++ core/src/main/java/com/graphhopper/routing/ev/MaxSlope.java
@@ -1,0 +1,12 @@
+package com.graphhopper.routing.ev;
+
+/**
+ * Maximum elevation change in m/100m.
+ */
+public class MaxSlope {
+    public static final String KEY = "max_slope";
+
+    public static DecimalEncodedValue create() {
+        return new DecimalEncodedValueImpl(KEY, 5, 1, false);
+    }
+}